sourcemodule: Move remaining sourcedevice* files to subfolder device
