+++ zera-modules/sourcemodule/src/module/sourcemoduleprogram.cpp
@@ -8,8 +8,8 @@
 #include "basedspmeasprogram.h"
 #include "sourcemoduleprogram.h"
 #include "sourcemodule.h"
-#include "../sourcedevicemanager.h"
-#include "../sourcedevicevein.h"
+#include "../device/sourcedevicemanager.h"
+#include "../device/sourcedevicevein.h"
 #include "../sourceinterface.h"
 
 SourceModuleProgram::SourceModuleProgram(SourceModule* module, std::shared_ptr<cBaseModuleConfiguration> pConfiguration) :
